pyproject.toml
[project]
name = "test-avantsoft-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Devis",email = "davi.ribeirodosanjos@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-filter (>=25.1,<26.0)",
    "dj-database-url (>=3.0.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
pytest-factoryboy = "^2.8.1"
model-bakery = "^1.20.5"

